% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomialReg.R
\name{geem_betabinomial}
\alias{geem_betabinomial}
\title{Fit a beta-binomial generalized estimating equation model.}
\usage{
geem_betabinomial(formula, N, id, waves, data = parent.frame(),
  corstr = "independence", weights = NULL, corr.mat = NULL,
  init.beta = NULL, init.alpha = NULL, init.phi = 1, init.rho = 0,
  rho.fixed = FALSE, scale.fix = FALSE, nodummy = FALSE,
  sandwich = TRUE, maxit = 20, tol = 1e-05)
}
\arguments{
\item{formula}{See corresponding documentation to \code{geem}.}

\item{N}{A vector of total counts of length \code{reponse}.}

\item{id}{See corresponding documentation to \code{geem}. id must be a factor or integer variable.}

\item{data}{See corresponding documentation to \code{geem}.}

\item{corstr}{See corresponding documentation to \code{geem}.}

\item{weights}{See corresponding documentation to \code{geem}.}

\item{corr.mat}{See corresponding documentation to \code{geem}.}

\item{init.beta}{See corresponding documentation to \code{geem}.}

\item{init.alpha}{See corresponding documentation to \code{geem}.}

\item{init.phi}{See corresponding documentation to \code{geem}.}

\item{scale.fix}{See corresponding documentation to \code{geem}.}

\item{nodummy}{See corresponding documentation to \code{geem}.}

\item{sandwich}{See corresponding documentation to \code{geem}.}

\item{maxit}{See corresponding documentation to \code{geem}.}

\item{tol}{See corresponding documentation to \code{geem}.}

\item{Mv}{See corresponding documentation to \code{geem}.}
}
\value{
An object of class "geem_betabinomial
}
\description{
Calculate coefficients and nuisance parameters for the beta-binomial
model using generalized estimating equations. Implementation is based
on the geeM package.
}

