% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomialReg.R
\name{predictBB}
\alias{predictBB}
\title{Prediction for Beta-Binomial regression.}
\usage{
predictBB(fit, N, X = NULL)
}
\arguments{
\item{fit}{A fitted geem_betabinomial object.}

\item{X}{An optional matrix of explanatory variables.}

\item{M}{A vector of total cell counts.}
}
\value{
Returns a vector of expected counts
}
\description{
Prediction for Beta-Binomial regression.
}

